feat(messages): display emoji-only DMs larger without bubble
Emoji-only messages now render WhatsApp-style: large text (text-4xl)
without the bubble background. Extracts isEmojiOnly to shared utils
for reuse across posts and DMs.
